% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Bar Chart}
\usage{
bar_chart(
  data,
  x,
  y,
  fill_var = NULL,
  label_var = NULL,
  label_color = "black",
  label_size = 3,
  title = "",
  fill = "#4E79A7",
  orientation = "h",
  position = "dodge",
  scale_axis_pct = FALSE,
  legend_position = "right"
)
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{x}{A categorical variable that represents the x-axis.}

\item{y}{A numeric variable that represents the y-axis.}

\item{fill_var}{(Optional) A categorical variable to group and fill the plot.}

\item{label_var}{(Optional) A variable used to label the bars.}

\item{label_color}{A string indicating the color of the label text when label_var is provide. (Default: "black").}

\item{label_size}{An integer specifying the size of the label text when label_var is provide. (Default: 3).}

\item{title}{A string indicating the title.}

\item{fill}{A string indicating the color to fill bars when fill_var is not provided. (Default: "#4E79A7").}

\item{orientation}{A string indicating the orientation of the chart. Can be "h" for horizontal or "v" for vertical. (Default "h").}

\item{position}{A string indicating the position of the bars. Can be "dodge", "stack", or "fill". (Default "dodge").}

\item{scale_axis_pct}{A boolean value indicating whether the y-axis should be scaled by percent. (Default FALSE).}

\item{legend_position}{A string indicating the position of the legend. Can be "right", "left", "top", or "bottom". (Default "right").}
}
\value{
A ggplot object
}
\description{
Plot a Bar Chart in ggplot2 with the option to fill by a categorical variable
}
\examples{
library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
  mutate(across(c(am, carb, cyl, gear, vs), as.factor))

data1_mtcars <- mtcars \%>\%
  count(cyl) \%>\%
 mutate(
   pct = round(n/sum(n),3),
   label = paste(pct*100, '\%')
   )

bar_chart(data1_mtcars, cyl, pct)

bar_chart(data1_mtcars, cyl, pct,
         fill_var = cyl,
         scale_axis_pct = TRUE,
         label_var = label,
         orientation = "v")


data2_mtcars <- mtcars \%>\%
  count(am, cyl) \%>\%
 group_by(am) \%>\%
 mutate(
   pct = round(n/sum(n),3),
   label = paste(pct*100, '\%')
   )

bar_chart(data2_mtcars, am, pct,
         label_var = label,
         scale_axis_pct = TRUE,
         label_color = "white",
         fill_var = cyl,
         orientation = "h",
         position = "stack")

}
