% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_chart.R
\name{bubble_chart}
\alias{bubble_chart}
\title{Bubble Chart}
\usage{
bubble_chart(
  data,
  x,
  y,
  size,
  fill,
  scale_size_range = c(0.5, 20),
  alpha_point = 0.8,
  color_point = "black",
  stroke_point = 0.2,
  legend_position = "right",
  title = "",
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{x}{A numeric variable.}

\item{y}{A numeric variable.}

\item{size}{A numeric variable indicating the size of the bubbles.}

\item{fill}{A variable that controls the fill color of the bubbles.}

\item{scale_size_range}{A numeric vector with the minimum and maximum size of the bubbles.}

\item{alpha_point}{A numeric value indicating the transparency of the bubbles.}

\item{color_point}{A character string indicating the color of the points.}

\item{stroke_point}{A numeric value indicating the width of the stroke of the points.}

\item{legend_position}{A character string indicating the position of the legend in the plot.}

\item{title}{A character string indicating the title of the plot.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Plot Bubble Chart in ggplot2
}
\examples{
library(dplyr)
library(gapminder)

data <- gapminder \%>\% filter(year=="2002") \%>\%
  dplyr::select(-year)

data_bubble <- data \%>\%
  mutate(pop=pop/1000000) \%>\%
  arrange(desc(pop)) \%>\%
  mutate(country = factor(country, country)) \%>\%
  mutate(
    annotation = case_when(
      gdpPercap > 5000 & lifeExp < 60 ~ "yes",
      lifeExp < 30 ~ "yes",
      gdpPercap > 40000 ~ "yes"
    )
  )


bubble_chart(
  data = data_bubble,
  x = gdpPercap,
  y = lifeExp,
  size = pop,
  fill = continent
)

bubble_chart(
  data = data_bubble,
  x = gdpPercap,
  y = lifeExp,
  size = pop,
  fill = continent
)+
  ggrepel::geom_text_repel(
    data = data_bubble \%>\% filter(annotation=="yes"),
    ggplot2::aes(label = country),
    size = 3
  )

}
