% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlation coefficient}
\usage{
correlate(.data, method = "pearson")
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{method}{a character string: "pearson", "kendall" or "spearman".}
}
\value{
A \code{tibble}
}
\description{
Correlation coefficient
}
\examples{

iris \%>\%
   correlate()


}
