% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveStats.R
\name{descriptiveStats}
\alias{descriptiveStats}
\title{Descriptive Statistics}
\usage{
descriptiveStats(.data)
}
\arguments{
\item{.data}{.data A \code{tibble} or \code{data.frame}}
}
\value{
A \code{tibble}
}
\description{
Descriptive Statistics
}
\examples{

library(dplyr)

iris \%>\%
   descriptiveStats()

iris \%>\%
   group_by(Species) \%>\%
   descriptiveStats()


}
