% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreCategory.R
\name{exploreCategory}
\alias{exploreCategory}
\title{Explore categorical variables}
\usage{
exploreCategory(.data, top = 30)
}
\arguments{
\item{.data}{a data frame or tibble}

\item{top}{an integer indicating the number of top levels to display (default is 30)}
}
\value{
A \code{tibble}
}
\description{
This function explores all categorical variables calculating the frequency and ratio of each level for each variable.
The output includes the top levels for each variable by frequency.
}
\examples{
iris \%>\%
   exploreCategory()

}
