% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{Explore a data frame or tibble}
\usage{
exploreData(.data)
}
\arguments{
\item{.data}{A data frame or tibble to summarize.}
}
\value{
A data frame with the following columns: column's data type, number of non-missing values,
number of missing values, number of unique values, and the percentage of unique values.
}
\description{
This function explores a all variables in a data frame.
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris <- as_tibble(iris)

iris \%>\%
   exploreData()

mtcars \%>\%
   exploreData()

}
