% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreNumeric.R
\name{exploreNumeric}
\alias{exploreNumeric}
\title{Summarize Numeric Data}
\usage{
exploreNumeric(.data)
}
\arguments{
\item{.data}{a data frame or tibble}
}
\value{
a tibble with one row for each numeric variable and columns for count, missing values,
minimum, maximum, median, quartiles, mean, standard deviation, number of zeros, number of negatives,
and number of outliers.
}
\description{
This function summary numeric variables including count, missing values, minimum, maximum, median, quartiles, mean, standard deviation,
number of zeros, number of negatives, and number of outliers.
}
\examples{
library(dplyr)

iris \%>\%
   exploreNumeric()

iris \%>\%
   select(Sepal.Length, Sepal.Width) \%>\%
   exploreNumeric()

iris \%>\%
  group_by(Species) \%>\%
  exploreNumeric()

mtcars <- mtcars \%>\% as_tibble() \%>\%
  mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\%
   dplyr::group_by(cyl, am) \%>\%
   exploreNumeric()

}
