% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreOutliers.R
\name{exploreOutliers}
\alias{exploreOutliers}
\title{Explore Outliers}
\usage{
exploreOutliers(.data)
}
\arguments{
\item{.data}{A data frame or tibble containing numeric columns.}
}
\value{
A data frame with the following columns: The name of the numeric variable,
number of non-missing observations, number of missing observations,
number of outliers detected using the boxplot method,
ratio of outliers to the total number of observations, The mean value of the outliers,
mean value of the variable including outliers, mean value of the variable after removing outliers.
}
\description{
This function explores the outliers of all numeric variables in a data frame.
}
\examples{

library(dplyr)

iris \%>\%
   exploreOutliers()

iris \%>\%
   group_by(Species) \%>\%
   exploreOutliers()

}
