% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  data,
  x,
  y,
  color_var = NULL,
  color = "#4E79A7",
  label_var = NULL,
  facet_wrap_var = NULL,
  facet_ncol = 2,
  scales = "free_y",
  title = "Plot Time Series",
  linewidth = 1,
  legend_position = "none",
  highlight = NULL,
  compare = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{A numeric variable}

\item{y}{A numeric variable}

\item{color_var}{(Optional) A categorical variable to group and color the plot.}

\item{color}{A string indicating the color to use for the line when color_var is not provide.}

\item{label_var}{(Optional) A categorical variable to use for labeling the lines.}

\item{facet_wrap_var}{(Optional) A categorical variable to use for faceting the plot.}

\item{facet_ncol}{An integer indicating the number of columns to use for facet_wrap. Defaults to 2.}

\item{scales}{A string indicating the type of scaling to use in facet (default: "free").}

\item{title}{A string indicating the title.}

\item{linewidth}{An integer indicating the width of the line.}

\item{legend_position}{A string indicating the position of the legend. Use "none" to hide the legend.}

\item{highlight}{A string or vector of characters to be highlighted in the graph when color_var is provided}

\item{compare}{a boolean indicating whether you want to compare the line with the others in gray colors when color_var and facet_wrap_var are provided}
}
\value{
A ggplot object
}
\description{
Plot Line Chart in ggplot2
}
\examples{

library(dplyr)

line_chart(ggplot2::economics, x = date, y = unemploy)

economics_long2 <-
  dplyr::filter(
    ggplot2::economics_long,
    variable \%in\% c("pop", "uempmed", "unemploy")
  )

line_chart(economics_long2, x = date, y = value01, color_var = variable)

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           highlight = "unemploy")

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_wrap_var = variable)

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_wrap_var = variable,  highlight = "unemploy",  scales = "fixed")

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_wrap_var = variable, compare = TRUE)

}
