% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Bar Plot based on target variable}
\usage{
plotBar(
  .data,
  interactive = FALSE,
  top = 10,
  pct = TRUE,
  position_group = "fill",
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{top}{Top}

\item{pct}{Pct}

\item{position_group}{Position}

\item{ncol_facet}{Number of facet columns}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Bar Plot based on target variable
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\%
   plotBar()

mtcars \%>\%
   group_by(cyl) \%>\%
   plotBar()

}
