% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelate.R
\name{plotCorrelate}
\alias{plotCorrelate}
\title{Visualize correlation coefficient}
\usage{
plotCorrelate(.data, method = "pearson", interactive = FALSE)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{method}{a character string: "pearson", "kendall" or "spearman".}

\item{interactive}{TRUE/FLASE. Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Visualize correlation coefficient
}
\examples{

library(dplyr)

iris \%>\%
   plotCorrelate(interactive = TRUE)


}
