% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot Density Chart in ggplot2}
\usage{
plotDensity(
  .data,
  fill_var = NULL,
  interactive = FALSE,
  facet_grid = FALSE,
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{fill_var}{(Optional) A categorical variable to group and fill the plot.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}

\item{facet_grid}{A boolean value that specifies whether the chart should be displayed in facet_grid when there are only two numeric variables and one categorical variable specified in fill_var.}

\item{ncol_facet}{An integer that specifies the number of columns in the facet plot. Default is 3.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
This function plots a Density Chart for all numeric variables,
with the option to fill by a categorical variable.
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris \%>\%
   plotDensity()

mtcars \%>\%
   plotDensity()

iris \%>\%
   plotDensity(fill_var = Species)

iris \%>\%
   select(Sepal.Width, Species) \%>\%
   plotDensity(fill_var = Species, facet_grid = TRUE)

}
