% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity_target.R
\name{plotDensity_target}
\alias{plotDensity_target}
\title{Visualize Distributions based on target variable and categorical variables}
\usage{
plotDensity_target(
  .data,
  target_num,
  dir_cat = "h",
  ncol_cat = 4,
  interactive = FALSE
)
}
\arguments{
\item{.data}{A data frame.}

\item{target_num}{The name of the target numeric variable to analyze.}

\item{dir_cat}{The direction of facet_wrap. Options are "h" for horizontal and "v" for vertical.}

\item{ncol_cat}{The number of columns in facet_wrap.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualize Distributions based on target variable and categorical variables
}
\examples{
library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\% select(mpg, cyl, gear) \%>\%
    plotDensity_target(target_num = mpg)

iris \%>\%
    plotDensity_target(target_num = Petal.Length)

}
