% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Title}
\usage{
plotHistogram(
  .data,
  interactive = FALSE,
  facet_grid = FALSE,
  ncol_facet = 3,
  bins = 30,
  alpha = 1,
  alpha_w_group = 0.7,
  scales = "free"
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{facet_grid}{facet grid}

\item{ncol_facet}{Number of facet columns}

\item{bins}{Number of bins}

\item{alpha}{alpha}

\item{alpha_w_group}{alpha f}

\item{scales}{scales "free", "free_x", "free_y"}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Title
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris \%>\%
   plotHistogram()

mtcars \%>\%
   plotHistogram()

}
