% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly-themes.R
\name{plotly_dark}
\alias{plotly_dark}
\alias{plotly_white}
\title{Plotly Themes}
\usage{
plotly_dark(
  plot,
  config.scrollZoom = FALSE,
  config.displayModeBar = FALSE,
  config.responsive = TRUE,
  config.staticPlot = FALSE,
  layout.dragmode = TRUE,
  layout.autosize = TRUE,
  layout.width = NULL,
  font.color = "#ffffff",
  font.family = "Segoe UI",
  layout.background = "#292c30",
  ...
)

plotly_white(
  plot,
  config.scrollZoom = FALSE,
  config.displayModeBar = FALSE,
  config.responsive = TRUE,
  config.staticPlot = FALSE,
  layout.dragmode = TRUE,
  layout.autosize = TRUE,
  layout.width = NULL,
  font.color = "#1c1c1c",
  font.family = "Segoe UI",
  layout.background = "#ffffff",
  ...
)
}
\arguments{
\item{plot}{A plotly object}

\item{config.scrollZoom}{logical. FALSE default}

\item{config.displayModeBar}{logical. FALSE default}

\item{config.responsive}{logical. TRUE default}

\item{config.staticPlot}{logical. FALSE default}

\item{layout.dragmode}{logical. TRUE default}

\item{layout.autosize}{logical. TRUE default}

\item{layout.width}{A number. NULL default}

\item{font.color}{A string. "#ffffff" default}

\item{font.family}{A string. "Segoe UI" default}

\item{layout.background}{A string. "#292c30" default}

\item{...}{Parameters passed to \code{plotly} functions.}
}
\value{
A plotly object
}
\description{
Plotly Themes
}
\examples{


library(dplyr)
library(plotly)

iris \%>\%
  plot_ly() \%>\%
  add_trace(x = ~Sepal.Width, y = ~Sepal.Length, color = ~Species,
            type = "scatter", mode = "markers") \%>\%
  plotly_dark() \%>\%
  layout(
    title = list(text = "<b>Iris Sepal.Length vs Sepal.Width<b>"),
    legend=list(title=list(text='Species'))
  )

}
