% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{Visualize missing data}
\usage{
plotMissing(.data, interactive = FALSE)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualize missing data
}
\examples{

iris \%>\%
   plotMissing()

}
