% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Plot Outliers}
\usage{
plotOutliers(.data, interactive = FALSE)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Plots a Density Chart and Box-Plot of numerical variables
}
\examples{
library(dplyr)
iris \%>\%
   select(Sepal.Width) \%>\%
   plotOutliers()

}
