% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaincloud.R
\name{plotRaincloud}
\alias{plotRaincloud}
\title{Visualize Distributions with Raincloud Plots}
\usage{
plotRaincloud(
  .data,
  interactive = FALSE,
  fill_ind = "#4e79a7",
  color_ind = "#4e79a7"
)
}
\arguments{
\item{.data}{.data A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{fill_ind}{fill}

\item{color_ind}{color}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualize Distributions with Raincloud Plots
}
\examples{

library(dplyr)

iris \%>\%
   group_by(Species) \%>\%
   select(Sepal.Width) \%>\%
   plotRaincloud()

}
