% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Scatter Plot based on target variable}
\usage{
plotScatter(
  .data,
  y,
  fill_var = NULL,
  smooth = TRUE,
  interactive = FALSE,
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{y}{A variable numeric}

\item{fill_var}{A categorical column that can be used to change the color}

\item{smooth}{smooth}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{ncol_facet}{Number of facet columns}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Scatter Plot based on target variable
}
\examples{

library(dplyr)

mtcars \%>\%
   plotScatter(y = disp)

iris \%>\%
   select(Petal.Width, Sepal.Length, Species) \%>\%
   plotScatter(y = Petal.Width, fill_var = Species)

}
