% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlations}
\usage{
correlate(.data, method = "pearson")
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{method}{The correlation method to use. One of "pearson", "kendall", or "spearman".}
}
\value{
A tibble with the pairwise correlations between numeric columns of the input data frame.
}
\description{
This function calculate pairwise correlations between numeric columns of a data frame.
}
\examples{

iris \%>\%
   correlate()


}
