% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BoxPlot_Outlier.R
\name{data_BoxPlot_Outlier}
\alias{data_BoxPlot_Outlier}
\title{Statistical Data: Box Plot}
\usage{
data_BoxPlot_Outlier(.data)
}
\arguments{
\item{.data}{.data A \code{tibble} or \code{data.frame}}
}
\value{
A \code{tibble}
}
\description{
Statistical Data: Box Plot
}
\examples{
library(dplyr)

iris \%>\% select(Sepal.Width) \%>\%
  data_BoxPlot_Outlier()

}
