% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveStats.R
\name{descriptiveStats}
\alias{descriptiveStats}
\title{Descriptive Statistic}
\usage{
descriptiveStats(.data)
}
\arguments{
\item{.data}{A data.frame or tibble.}
}
\value{
A tibble with descriptive statistics for all numeric variables in .data.
}
\description{
This function calculate descriptive statistics for numeric variables in a data frame
}
\examples{

library(dplyr)

iris \%>\%
   descriptiveStats()

iris \%>\%
   group_by(Species) \%>\%
   descriptiveStats()


}
