% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreCategory.R
\name{exploreCategory}
\alias{exploreCategory}
\title{Description of all categorical variables}
\usage{
exploreCategory(.data, top = 500)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{top}{Number of levels to return for each variable}
}
\value{
A \code{tibble}
}
\description{
Description of all categorical variables
}
\examples{
iris \%>\%
   exploreCategory()

}
