% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{General description of all variables}
\usage{
exploreData(.data)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}
}
\value{
A \code{tibble}
}
\description{
General description of all variables
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris <- as_tibble(iris)

iris \%>\%
   exploreData()

mtcars \%>\%
   exploreData()

}
