% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreNumeric.R
\name{exploreNumeric}
\alias{exploreNumeric}
\title{Description of all numerical variables}
\usage{
exploreNumeric(.data)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}
}
\value{
A \code{tibble}
}
\description{
Description of all numerical variables
}
\examples{
library(dplyr)

iris \%>\%
   exploreNumeric()

iris \%>\%
   select(Sepal.Length, Sepal.Width) \%>\%
   exploreNumeric()

iris \%>\%
  group_by(Species) \%>\%
  exploreNumeric()

mtcars <- mtcars \%>\% as_tibble() \%>\%
  mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\%
   dplyr::group_by(cyl, am) \%>\%
   exploreNumeric()

}
