% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreOutliers.R
\name{exploreOutliers}
\alias{exploreOutliers}
\title{Describe the outliers of the numeric variables of the data frame}
\usage{
exploreOutliers(.data)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}
}
\value{
A \code{tibble}
}
\description{
Describe the outliers of the numeric variables of the data frame
}
\examples{

library(dplyr)

iris \%>\%
   exploreOutliers()

iris \%>\%
   group_by(Species) \%>\%
   exploreOutliers()

}
