% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  data,
  x,
  y,
  color_var = NULL,
  color = "#4E79A7",
  label_var = NULL,
  facet_wrap_var = NULL,
  facet_ncol = 2,
  scales = "free",
  linewidth = 1
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{x}{A numeric variable}

\item{y}{A numeric variable}

\item{color_var}{(Optional) A categorical variable to group and color the plot.}

\item{color}{A string indicating the color to use for the line when color_var is not provide.}

\item{label_var}{(Optional) A categorical variable to use for labeling the lines.}

\item{facet_wrap_var}{(Optional) A categorical variable to use for faceting the plot.}

\item{facet_ncol}{An integer indicating the number of columns to use for facet_wrap. Defaults to 2.}

\item{scales}{A string indicating the type of scaling to use in facet (default: "free").}

\item{linewidth}{The width of the line.}
}
\value{
A ggplot object
}
\description{
Plot Line Chart in ggplot2
}
\examples{

library(dplyr)
library(gapminder)

line_chart(ggplot2::economics,
           x = date,
           y = unemploy)

economics_long2 <-
  dplyr::filter(
    ggplot2::economics_long,
    variable \%in\% c("pop", "uempmed", "unemploy")
  )

line_chart(economics_long2,
           x = date,
           y = value01,
           color_var = variable)

line_chart(economics_long2,
           x = date,
           y = value01,
           color_var = variable,
           facet_wrap_var = variable,
           facet_ncol = 2)


multiple_line_df <- gapminder::gapminder \%>\%
  filter(country \%in\% c("China", "United States", "Ecuador", "Peru")) \%>\%
  mutate(label = if_else(year == max(year), as.character(country), NA_character_))


line_chart(multiple_line_df,
           x = year,
           y = lifeExp,
           color_var = country)

line_chart(multiple_line_df,
           x = year,
           y = lifeExp,
           color_var = country,
           facet_wrap_var = country)

}
