% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Plot bar chart for categorical variables}
\usage{
plotBar(
  .data,
  fill_var = NULL,
  interactive = FALSE,
  top = 10,
  pct = TRUE,
  position_group = "fill",
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{fill_var}{(Optional) A categorical variable to group and fill the plot.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}

\item{top}{An integer specifying the number of top variables to display. Defaults to 10.}

\item{pct}{A boolean value indicating whether to plot frequencies or proportions. Defaults to TRUE.}

\item{position_group}{A string indicating the position of the bars. Can be "fill" or "dodge". Defaults to "fill".}

\item{ncol_facet}{An integer indicating the number of columns to use for facet_wrap. Defaults to 3.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
This function plots a Bar Chart for all categorical variables, with the option to fill by a categorical variable.
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\%
   plotBar()

mtcars \%>\%
   plotBar(fill_var = cyl, position_group = "fill")

}
