% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelate.R
\name{plotCorrelate}
\alias{plotCorrelate}
\title{Plot Correlation Coefficient}
\usage{
plotCorrelate(.data, method = "pearson", interactive = FALSE)
}
\arguments{
\item{.data}{A data frame or tibble with numeric columns to calculate correlations.}

\item{method}{The correlation coefficient method to use. Defaults to "pearson".}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
A function plots of the correlation coefficients between all pairs of numeric variables in a data frame.
}
\examples{

library(dplyr)

iris \%>\%
   plotCorrelate(interactive = TRUE)


}
