% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Visualize Distributions}
\usage{
plotDensity(
  .data,
  fill_var = NULL,
  interactive = FALSE,
  facet_grid = FALSE,
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{fill_var}{A categorical column that can be used to change the color}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{facet_grid}{facet}

\item{ncol_facet}{Number of facet columns}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualize Distributions
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris \%>\%
   plotDensity()

mtcars \%>\%
   plotDensity()

iris \%>\%
   plotDensity(fill_var = Species)

iris \%>\%
   select(Sepal.Width, Species) \%>\%
   plotDensity(fill_var = Species, facet_grid = TRUE)

}
