% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity_target.R
\name{plotDensity_target}
\alias{plotDensity_target}
\title{Visualize Distributions based on target variable and categorical variables}
\usage{
plotDensity_target(
  .data,
  target_num,
  dir_cat = "h",
  ncol_cat = 4,
  interactive = FALSE
)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame} with categorical variables and target numerical variable}

\item{target_num}{A variable numeric}

\item{dir_cat}{Direction of facet columns}

\item{ncol_cat}{Number of facet columns}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Visualize Distributions based on target variable and categorical variables
}
\examples{
library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\% select(mpg, cyl, gear) \%>\%
    plotDensity_target(target_num = mpg)

iris \%>\%
    plotDensity_target(target_num = Petal.Length)

}
