% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMissing.R
\name{plotMissing}
\alias{plotMissing}
\title{Plot the Percentage of Missing Values in a Data Frame}
\usage{
plotMissing(.data, interactive = FALSE)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}
}
\description{
This function plots the percentage of missing values for each variable in the data frame.
The output is a bar chart with the variables sorted in descending order based on the percentage of missing values.
}
\examples{

iris \%>\%
   plotMissing()

}
