% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormality.R
\name{plotNormality}
\alias{plotNormality}
\title{Visualization of normality}
\usage{
plotNormality(
  .data,
  interactive = FALSE,
  color_density = "steelblue",
  color_qq = "steelblue",
  color_qq_line = "steelblue4",
  color_sqrt = "darkorange",
  color_log = "salmon"
)
}
\arguments{
\item{.data}{.data  A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}

\item{color_density}{Density Color}

\item{color_qq}{QQ Color}

\item{color_qq_line}{QQ Line Color}

\item{color_sqrt}{Sqrt Color}

\item{color_log}{Log Color}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualization of normality
}
\examples{

library(dplyr)

iris \%>\%
   select(Petal.Width) \%>\%
   plotNormality()

iris \%>\%
   select(Petal.Width, Sepal.Length) \%>\%
   plotNormality()

}
