% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Visualization of outliers}
\usage{
plotOutliers(.data, interactive = FALSE)
}
\arguments{
\item{.data}{A \code{tibble} or \code{data.frame}}

\item{interactive}{Returns either a static (\code{ggplot2}) visualization or an interactive (\code{plotly}) visualization}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Visualization of outliers
}
\examples{
library(dplyr)
iris \%>\%
   select(Sepal.Width) \%>\%
   plotOutliers()

}
