% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Plot Scatter Chart in ggplot2}
\usage{
plotScatter(
  .data,
  y,
  fill_var = NULL,
  smooth = TRUE,
  interactive = FALSE,
  ncol_facet = 3
)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{y}{A target numeric variable to be plotted on the y-axis.}

\item{fill_var}{(Optional) A categorical variable to group and fill the plot.}

\item{smooth}{A boolean that specifies whether a linear regression line should be added to the plot. Default is TRUE.}

\item{interactive}{A boolean that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}

\item{ncol_facet}{A Boolean that specifies number of columns for the facet_wrap function. Default is 3.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
This function plots a Scatter Chart for all numeric variables and a target numeric variable (y-axis),
with the option to fill by a categorical variable.
}
\examples{

library(dplyr)

mtcars \%>\%
   plotScatter(y = disp)

iris \%>\%
   select(Petal.Width, Sepal.Length, Species) \%>\%
   plotScatter(y = Petal.Width, fill_var = Species)

}
